chore: moved page classes in pages repository
