+++ test/src/test/java/com/test/urubupix/pages/TransactionPage.java
@@ -1,7 +1,5 @@
-package com.test.urubupix;
+package com.test.urubupix.pages;
 
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 
 public class TransactionPage {
     private WebDriver driver;
